refact code, well done
