+++ maze/.idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/testMTS.iml" filepath="$PROJECT_DIR$/.idea/testMTS.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/maze.iml" filepath="$PROJECT_DIR$/.idea/maze.iml" />
     </modules>
   </component>
 </project>
