--- maze/Makefile.txt
@@ -1,14 +1,0 @@
-TARGET = main.out
-HDRS_DIR = testMTS
-
-SRCS = \
-       project/main.c \
-       project/func.c
-
-.PHONY: all clean
-
-all: $(SRCS)
-	$(CC) -Wall -Wextra -Werror -I $(HDRS_DIR) -o $(TARGET) $(CFLAGS) $(SRCS)
-
-clean:
-	rm -rf $(TARGET)
