+++ maze/test.sh
