--- maze/testMTS/CMakeLists.txt
@@ -1,6 +1,0 @@
-cmake_minimum_required(VERSION 3.14)
-project(testMTS)
-
-set(CMAKE_CXX_STANDARD 14)
-
-add_executable(testMTS main.cpp func.cpp func.h)
