--- maze/testMTS/cmake-build-debug/CMakeFiles/Makefile.cmake
@@ -1,52 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.14
-
-# The generator used is:
-set(CMAKE_DEPENDS_GENERATOR "Unix Makefiles")
-
-# The top level Makefile was generated from the following files:
-set(CMAKE_MAKEFILE_DEPENDS
-  "CMakeCache.txt"
-  "../CMakeLists.txt"
-  "CMakeFiles/3.14.3/CMakeCCompiler.cmake"
-  "CMakeFiles/3.14.3/CMakeCXXCompiler.cmake"
-  "CMakeFiles/3.14.3/CMakeSystem.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/CMakeCInformation.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/CMakeCXXInformation.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/CMakeCheckCompilerFlagCommonPatterns.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/CMakeCommonLanguageInclude.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/CMakeExtraGeneratorDetermineCompilerMacrosAndIncludeDirs.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/CMakeFindCodeBlocks.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/CMakeGenericSystem.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/CMakeInitializeConfigs.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/CMakeLanguageInformation.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/CMakeSystemSpecificInformation.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/CMakeSystemSpecificInitialize.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/Compiler/CMakeCommonCompilerMacros.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/Compiler/GNU-C.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/Compiler/GNU-CXX.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/Compiler/GNU.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/Internal/CMakeCheckCompilerFlag.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/Platform/Linux-GNU-C.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/Platform/Linux-GNU-CXX.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/Platform/Linux-GNU.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/Platform/Linux.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/Platform/UnixPaths.cmake"
-  "/home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/share/cmake-3.14/Modules/ProcessorCount.cmake"
-  )
-
-# The corresponding makefile is:
-set(CMAKE_MAKEFILE_OUTPUTS
-  "Makefile"
-  "CMakeFiles/cmake.check_cache"
-  )
-
-# Byproducts of CMake generate step:
-set(CMAKE_MAKEFILE_PRODUCTS
-  "CMakeFiles/CMakeDirectoryInformation.cmake"
-  )
-
-# Dependency information for all targets:
-set(CMAKE_DEPEND_INFO_FILES
-  "CMakeFiles/testMTS.dir/DependInfo.cmake"
-  )
