--- maze/testMTS/cmake-build-debug/CMakeFiles/TargetDirectories.txt
@@ -1,3 +1,0 @@
-/home/sancho/CLionProjects/testMTS/cmake-build-debug/CMakeFiles/rebuild_cache.dir
-/home/sancho/CLionProjects/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir
-/home/sancho/CLionProjects/testMTS/cmake-build-debug/CMakeFiles/edit_cache.dir
