--- maze/testMTS/cmake-build-debug/CMakeFiles/clion-environment.txt
@@ -1,3 +1,0 @@
-ToolSet: 1.0 (local)Options: 
-
-Options:
