--- maze/testMTS/cmake-build-debug/CMakeFiles/progress.marks
@@ -1,1 +1,0 @@
-3
