--- maze/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir/CXX.includecache
@@ -1,32 +1,0 @@
-#IncludeRegexLine: ^[ 	]*[#%][ 	]*(include|import)[ 	]*[<"]([^">]+)([">])
-
-#IncludeRegexScan: ^.*$
-
-#IncludeRegexComplain: ^$
-
-#IncludeRegexTransform: 
-
-/home/sancho/CLionProjects/testMTS/func.h
-iostream
--
-fstream
--
-errno.h
--
-vector
--
-queue
--
-
-/home/sancho/CLionProjects/testMTS/main.cpp
-iostream
--
-fstream
--
-vector
--
-queue
--
-func.h
-/home/sancho/CLionProjects/testMTS/func.h
-
