--- maze/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir/DependInfo.cmake
@@ -1,21 +1,0 @@
-# The set of languages for which implicit dependencies are needed:
-set(CMAKE_DEPENDS_LANGUAGES
-  "CXX"
-  )
-# The set of files for implicit dependencies of each language:
-set(CMAKE_DEPENDS_CHECK_CXX
-  "/home/sancho/CLionProjects/testMTS/func.cpp" "/home/sancho/CLionProjects/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir/func.cpp.o"
-  "/home/sancho/CLionProjects/testMTS/main.cpp" "/home/sancho/CLionProjects/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir/main.cpp.o"
-  )
-set(CMAKE_CXX_COMPILER_ID "GNU")
-
-# The include file search paths:
-set(CMAKE_CXX_TARGET_INCLUDE_PATH
-  )
-
-# Targets to which this target links.
-set(CMAKE_TARGET_LINKED_INFO_FILES
-  )
-
-# Fortran module output directory.
-set(CMAKE_Fortran_TARGET_MODULE_DIR "")
