--- maze/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir/build.make
@@ -1,113 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.14
-
-# Delete rule output on recipe failure.
-.DELETE_ON_ERROR:
-
-
-#=============================================================================
-# Special targets provided by cmake.
-
-# Disable implicit rules so canonical targets will work.
-.SUFFIXES:
-
-
-# Remove some rules from gmake that .SUFFIXES does not remove.
-SUFFIXES =
-
-.SUFFIXES: .hpux_make_needs_suffix_list
-
-
-# Suppress display of executed commands.
-$(VERBOSE).SILENT:
-
-
-# A target that is always out of date.
-cmake_force:
-
-.PHONY : cmake_force
-
-#=============================================================================
-# Set environment variables for the build.
-
-# The shell in which to execute make rules.
-SHELL = /bin/sh
-
-# The CMake executable.
-CMAKE_COMMAND = /home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/bin/cmake
-
-# The command to remove a file.
-RM = /home/sancho/Documents/clion-2019.1.3/bin/cmake/linux/bin/cmake -E remove -f
-
-# Escaping for special characters.
-EQUALS = =
-
-# The top-level source directory on which CMake was run.
-CMAKE_SOURCE_DIR = /home/sancho/CLionProjects/testMTS
-
-# The top-level build directory on which CMake was run.
-CMAKE_BINARY_DIR = /home/sancho/CLionProjects/testMTS/cmake-build-debug
-
-# Include any dependencies generated for this target.
-include CMakeFiles/testMTS.dir/depend.make
-
-# Include the progress variables for this target.
-include CMakeFiles/testMTS.dir/progress.make
-
-# Include the compile flags for this target's objects.
-include CMakeFiles/testMTS.dir/flags.make
-
-CMakeFiles/testMTS.dir/main.cpp.o: CMakeFiles/testMTS.dir/flags.make
-CMakeFiles/testMTS.dir/main.cpp.o: ../main.cpp
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/home/sancho/CLionProjects/testMTS/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_1) "Building CXX object CMakeFiles/testMTS.dir/main.cpp.o"
-	/usr/bin/c++  $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -o CMakeFiles/testMTS.dir/main.cpp.o -c /home/sancho/CLionProjects/testMTS/main.cpp
-
-CMakeFiles/testMTS.dir/main.cpp.i: cmake_force
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Preprocessing CXX source to CMakeFiles/testMTS.dir/main.cpp.i"
-	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -E /home/sancho/CLionProjects/testMTS/main.cpp > CMakeFiles/testMTS.dir/main.cpp.i
-
-CMakeFiles/testMTS.dir/main.cpp.s: cmake_force
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/testMTS.dir/main.cpp.s"
-	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S /home/sancho/CLionProjects/testMTS/main.cpp -o CMakeFiles/testMTS.dir/main.cpp.s
-
-CMakeFiles/testMTS.dir/func.cpp.o: CMakeFiles/testMTS.dir/flags.make
-CMakeFiles/testMTS.dir/func.cpp.o: ../func.cpp
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/home/sancho/CLionProjects/testMTS/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_2) "Building CXX object CMakeFiles/testMTS.dir/func.cpp.o"
-	/usr/bin/c++  $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -o CMakeFiles/testMTS.dir/func.cpp.o -c /home/sancho/CLionProjects/testMTS/func.cpp
-
-CMakeFiles/testMTS.dir/func.cpp.i: cmake_force
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Preprocessing CXX source to CMakeFiles/testMTS.dir/func.cpp.i"
-	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -E /home/sancho/CLionProjects/testMTS/func.cpp > CMakeFiles/testMTS.dir/func.cpp.i
-
-CMakeFiles/testMTS.dir/func.cpp.s: cmake_force
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/testMTS.dir/func.cpp.s"
-	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S /home/sancho/CLionProjects/testMTS/func.cpp -o CMakeFiles/testMTS.dir/func.cpp.s
-
-# Object files for target testMTS
-testMTS_OBJECTS = \
-"CMakeFiles/testMTS.dir/main.cpp.o" \
-"CMakeFiles/testMTS.dir/func.cpp.o"
-
-# External object files for target testMTS
-testMTS_EXTERNAL_OBJECTS =
-
-testMTS: CMakeFiles/testMTS.dir/main.cpp.o
-testMTS: CMakeFiles/testMTS.dir/func.cpp.o
-testMTS: CMakeFiles/testMTS.dir/build.make
-testMTS: CMakeFiles/testMTS.dir/link.txt
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/home/sancho/CLionProjects/testMTS/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_3) "Linking CXX executable testMTS"
-	$(CMAKE_COMMAND) -E cmake_link_script CMakeFiles/testMTS.dir/link.txt --verbose=$(VERBOSE)
-
-# Rule to build all files generated by this target.
-CMakeFiles/testMTS.dir/build: testMTS
-
-.PHONY : CMakeFiles/testMTS.dir/build
-
-CMakeFiles/testMTS.dir/clean:
-	$(CMAKE_COMMAND) -P CMakeFiles/testMTS.dir/cmake_clean.cmake
-.PHONY : CMakeFiles/testMTS.dir/clean
-
-CMakeFiles/testMTS.dir/depend:
-	cd /home/sancho/CLionProjects/testMTS/cmake-build-debug && $(CMAKE_COMMAND) -E cmake_depends "Unix Makefiles" /home/sancho/CLionProjects/testMTS /home/sancho/CLionProjects/testMTS /home/sancho/CLionProjects/testMTS/cmake-build-debug /home/sancho/CLionProjects/testMTS/cmake-build-debug /home/sancho/CLionProjects/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir/DependInfo.cmake --color=$(COLOR)
-.PHONY : CMakeFiles/testMTS.dir/depend
-
