--- maze/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir/cmake_clean.cmake
@@ -1,11 +1,0 @@
-file(REMOVE_RECURSE
-  "CMakeFiles/testMTS.dir/main.cpp.o"
-  "CMakeFiles/testMTS.dir/func.cpp.o"
-  "testMTS.pdb"
-  "testMTS"
-)
-
-# Per-language clean rules from dependency scanning.
-foreach(lang CXX)
-  include(CMakeFiles/testMTS.dir/cmake_clean_${lang}.cmake OPTIONAL)
-endforeach()
