--- maze/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir/depend.internal
@@ -1,9 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.14
-
-CMakeFiles/testMTS.dir/func.cpp.o
- /home/sancho/CLionProjects/testMTS/func.cpp
- /home/sancho/CLionProjects/testMTS/func.h
-CMakeFiles/testMTS.dir/main.cpp.o
- /home/sancho/CLionProjects/testMTS/func.h
- /home/sancho/CLionProjects/testMTS/main.cpp
