--- maze/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir/depend.make
@@ -1,9 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.14
-
-CMakeFiles/testMTS.dir/func.cpp.o: ../func.cpp
-CMakeFiles/testMTS.dir/func.cpp.o: ../func.h
-
-CMakeFiles/testMTS.dir/main.cpp.o: ../func.h
-CMakeFiles/testMTS.dir/main.cpp.o: ../main.cpp
-
