--- maze/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir/flags.make
@@ -1,10 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.14
-
-# compile CXX with /usr/bin/c++
-CXX_FLAGS = -g   -std=gnu++14
-
-CXX_DEFINES = 
-
-CXX_INCLUDES = 
-
