--- maze/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir/link.txt
@@ -1,1 +1,0 @@
-/usr/bin/c++  -g   CMakeFiles/testMTS.dir/main.cpp.o CMakeFiles/testMTS.dir/func.cpp.o  -o testMTS
