--- maze/testMTS/cmake-build-debug/CMakeFiles/testMTS.dir/progress.make
@@ -1,4 +1,0 @@
-CMAKE_PROGRESS_1 = 1
-CMAKE_PROGRESS_2 = 2
-CMAKE_PROGRESS_3 = 3
-
