--- maze/testMTS/cmake-build-debug/test.txt
@@ -1,10 +1,0 @@
-# ########
-#    #   #
-#### # # #
-#    #   #
-# ## ### #
-#    #s# #
-# #### # #
-#    # # #
-####     #
-##########
