--- maze/testMTS/func.h
@@ -1,22 +1,0 @@
-//
-// Created by sancho on 04.06.19.
-//
-
-#ifndef TESTMTS_FUNC_H
-#define TESTMTS_FUNC_H
-
-#include <iostream>
-#include <fstream>
-#include <errno.h>
-#include <vector>
-#include <queue>
-
-
-
-class func {
-public:
-    char GetMap(char* filename, std::vector <std::string> &v);
-};
-
-
-#endif //TESTMTS_FUNC_H
